Move all Ruby and SO files into lib/image_science
