+++ Rakefile
@@ -3,13 +3,13 @@ Bundler::GemHelper.install_tasks
 
 
 task :chmod do
-  File.chmod(0775, 'lib/image_science/image_science.so')
+  File.chmod(0775, 'lib/image_science/extension.so')
 end
 
 task :compile do
   `ruby ext/image_science/extconf.rb`
   `make`
-  `mv image_science.so lib/image_science/`
+  `mv image_science.so lib/image_science/extension.so`
 end
 
 task :build => [:compile, :chmod]
