+++ image_science.gemspec
@@ -1,5 +1,5 @@
 $:.push File.expand_path("../lib", __FILE__)
-require "image_science"
+require "image_science/version"
 
 Gem::Specification.new do |s|
   s.name = "image_science"
