+++ lib/image_science/version.rb
@@ -1,0 +1,3 @@
+class ImageScience
+  VERSION = '1.3.Asynchrony'
+end
